#14 Requested changes: Move logger.py to establishing folder
