+++ sosia/establishing/__init__.py
@@ -4,3 +4,4 @@ and accessing data in `sosia`.
 from sosia.establishing.constants import *
 from sosia.establishing.database import *
 from sosia.establishing.fields_sources import *
+from sosia.establishing.logger import *
